change delete user to auth
