--- src/Interfaces/IApiMarker.cs
@@ -1,6 +1,0 @@
-namespace SACA.Interfaces;
-
-public interface IApiMarker
-{
-    
-}
